refactor(workflow): extract budget bookkeeping
